+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/MediaAdapters/AudioMediaAdapter.swift
@@ -135,6 +135,10 @@ final class AudioMediaAdapter: MediaAdapting, @unchecked Sendable {
         try await localMediaManager.didUpdateCallSettings(settings)
     }
 
+    func didUpdateOwnCapabilities(_ ownCapabilities: Set<OwnCapability>) {
+        localMediaManager.didUpdateOwnCapabilities(ownCapabilities)
+    }
+
     /// Updates the publish options for the audio media adapter.
     ///
     /// - Parameter publishOptions: The new publish options to be applied.